feat(agents): add Gajae Code to the built-in agent registry
Gajae Code (gjc) is a terminal coding agent that runs on an existing Claude/Codex
subscription. It launches as a plain interactive TUI in the worktree cwd, so it
only needed registry wiring: detection, display name, telemetry kind, icon, and
auto-pick order.

The launch table moved to src/shared/tui-agent-config-table.ts because one more
agent pushed tui-agent-config.ts past the 300-line ceiling; per AGENTS.md the file
is split instead of baselined, and the table is re-exported from the old path so no
importer changes.
