+++ config/scripts/locale-translation-policy.mjs
@@ -43,6 +43,7 @@ export const NEVER_TRANSLATE_VALUES = new Set([
   'Cursor',
   'Droid',
   'Devin',
+  'Gajae Code',
   'Gemini',
   'Git',
   'Git Bash',
