+++ src/renderer/src/i18n/locales/en.json
@@ -420,6 +420,7 @@
           "8a9ba743cc": "Hermes",
           "4e63c7b956": "Rovo Dev",
           "bee242fe3d": "Qwen Code",
+          "c851f042e2": "Gajae Code",
           "ca73055bd0": "Mistral Vibe",
           "28810273af": "Kimi",
           "739a930554": "Droid",
