+++ src/renderer/src/lib/agent-favicon-assets.ts
@@ -24,6 +24,7 @@ import rovoUrl from '../../../shared/agent-icons/rovo.png?url'
 import hermesUrl from '../../../shared/agent-icons/hermes.png?url'
 import devinUrl from '../../../shared/agent-icons/devin.png?url'
 import openclawUrl from '../../../shared/agent-icons/openclaw.png?url'
+import gjcUrl from '../../../shared/agent-icons/gjc.png?url'
 
 // Why: these agents have no hand-authored SVG glyph, so previously their icons
 // loaded live from Google's favicon service. That service is unreachable in some
@@ -57,5 +58,6 @@ export const AGENT_FAVICON_ASSETS: Partial<Record<TuiAgent, string>> = {
   rovo: rovoUrl,
   hermes: hermesUrl,
   devin: devinUrl,
-  openclaw: openclawUrl
+  openclaw: openclawUrl,
+  gjc: gjcUrl
 }
