+++ src/shared/telemetry-events.ts
@@ -100,6 +100,7 @@ export const AGENT_KIND_VALUES = [
   'devin',
   'ante',
   'trae',
+  'gjc',
   'other'
 ] as const
 export const agentKindSchema = z.enum(AGENT_KIND_VALUES)
