+++ src/shared/tui-agent-selection.ts
@@ -17,6 +17,7 @@ export const TUI_AGENT_AUTO_PICK_ORDER = [
   'pi',
   'omp',
   'prime-agent',
+  'gjc',
   'gemini',
   'antigravity',
   'aider',
